backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
structlog==25.4.0
websockets==15.0.1
boto3==1.40.19
google-auth==2.40.3
huggingface-hub==0.34.4
pytest==8.4.1
pytest-asyncio==1.1.0
